Rename pack.
